ensemble/requirements.txt
# Ensemble — Python 3.11+
anthropic==0.39.0
ddgs>=3.0
elevenlabs==1.13.5
fastapi==0.115.6
httpx==0.28.1
python-dotenv==1.0.1
pyttsx3==2.90
uvicorn[standard]==0.34.0
